+++ example/anonymous-off-chain/voting.ts
@@ -1,7 +1,7 @@
 import * as Bluebird from "bluebird"
 import * as assert from "assert"
 import { INewProcessParams, ProcessMetadata, ProcessMetadataTemplate } from "@vocdoni/data-models"
-import { EthWaiter, VochainWaiter, Voting, VotingApi } from "@vocdoni/voting"
+import { AnonymousEnvelopeParams, EthWaiter, VochainWaiter, Voting, VotingApi, ProcessState } from "@vocdoni/voting"
 import {
   ProcessCensusOrigin,
   ProcessContractParameters,
@@ -11,10 +11,12 @@ import {
 } from "@vocdoni/contract-wrappers"
 import { getConfig } from "./config"
 import { TestAccount } from "./census"
-import { getChoicesForVoter } from "./util"
-import { CensusOffChainApi } from "@vocdoni/census"
+import { getChoicesForVoter, waitUntilPresent } from "./util"
+import { CensusOffChainApi, CensusOnChainApi, ZkInputs, ZkSnarks } from "@vocdoni/census"
 import { Wallet } from "@ethersproject/wallet"
 import { IGatewayClient } from "@vocdoni/client"
+import { Random } from "@vocdoni/common"
+import { Poseidon } from "@vocdoni/hashing"
 
 const config = getConfig()
 
@@ -43,12 +45,12 @@ export async function launchNewVote(censusRoot: string, censusUri: string, entit
 
   console.log("Getting the block height")
   const currentBlock = await VotingApi.getBlockHeight(gwPool)
-  const startBlock = currentBlock + 25
+  const startBlock = currentBlock + 15
   const blockCount = 60480
 
   const processParamsPre: INewProcessParams = {
-    mode: ProcessMode.make({ autoStart: true, interruptible: true }), // helper
-    envelopeType: ProcessEnvelopeType.ENCRYPTED_VOTES, // bit mask
+    mode: ProcessMode.make({ autoStart: true, interruptible: true, preregister: true }),
+    envelopeType: ProcessEnvelopeType.make({ encryptedVotes: false, anonymousVoters: true }),
     censusOrigin: ProcessCensusOrigin.OFF_CHAIN_TREE,
     metadata: processMetadataPre,
     censusRoot,
@@ -67,6 +69,9 @@ export async function launchNewVote(censusRoot: string, censusUri: string, entit
   const processId = await VotingApi.newProcess(processParamsPre, entityWallet, gwPool)
   assert(processId)
 
+  // Waiting a bit
+  await waitUntilPresent(processId, gwPool)
+
   // Reading back
   const processParams = await VotingApi.getProcessContractParameters(processId, gwPool)
   assert.strictEqual(processParams.entityAddress.toLowerCase(), entityWallet.address.toLowerCase())
@@ -80,31 +85,95 @@ export async function launchNewVote(censusRoot: string, censusUri: string, entit
   return { processId, processParams, processMetadata }
 }
 
-export async function submitVotes(processId: string, processParams: ProcessContractParameters, processMetadata: ProcessMetadata, accounts: TestAccount[], gwPool: IGatewayClient) {
+export async function registerVoterKeys(processId: string, processParams: ProcessContractParameters, accounts: TestAccount[], gwPool: IGatewayClient) {
+  console.log("Registering keys")
+
+  await Bluebird.map(accounts, async (account: TestAccount, idx: number) => {
+    process.stdout.write(`Registering [${idx}] ; `)
+
+    // The key (within the census) to sign the request
+    const wallet = new Wallet(account.privateKey)
+
+    // Generate the random secret key that will be used for voting
+    const secretKey = Random.getBigInt(Poseidon.Q)
+    account.secretKey = secretKey
+
+    // Get a census proof to be able to register the new key
+    const censusProof = await CensusOffChainApi.generateProof(processParams.censusRoot, { key: account.publicKeyEncoded }, gwPool)
+    const requestedWeight = censusProof.weight
+
+    const proof = Voting.packageSignedProof(processId, processParams.censusOrigin, censusProof)
+
+    return CensusOnChainApi.registerVoterKey(processId, proof, secretKey, requestedWeight, wallet, gwPool)
+  })
+}
+
+
+export async function submitVotes(processId: string, processParams: ProcessState, processMetadata: ProcessMetadata, accounts: TestAccount[], gwPool: IGatewayClient) {
   console.log("Launching votes")
 
-  const processKeys = processParams.envelopeType.hasEncryptedVotes ? await VotingApi.getProcessKeys(processId, gwPool) : null
+  const state = await VotingApi.getProcessState(processId, gwPool)
+  const circuitInfo = await VotingApi.getProcessCircuitInfo(processId, gwPool)
+  const witnessGeneratorWasm = await VotingApi.fetchAnonymousWitnessGenerator(circuitInfo)
+  const zKey = await VotingApi.fetchAnonymousVotingZKey(circuitInfo)
+  const vKey = await VotingApi.fetchAnonymousVotingVerificationKey(circuitInfo)
+  const processKeys = processParams.envelopeType.encryptedVotes ? await VotingApi.getProcessKeys(processId, gwPool) : null
 
   await Bluebird.map(accounts, async (account: TestAccount, idx: number) => {
 
     // VOTER
-    const wallet = new Wallet(account.privateKey)
-
-    const censusProof = await CensusOffChainApi.generateProof(processParams.censusRoot, { key: account.publicKeyEncoded }, true, gwPool)
+    const censusProof = await CensusOnChainApi.generateProof(state.rollingCensusRoot, account.secretKey, gwPool)
 
     const choices = getChoicesForVoter(processMetadata.questions.length, idx)
 
-    const envelope = processParams.envelopeType.hasEncryptedVotes ?
-      Voting.packageSignedEnvelope({ censusOrigin: processParams.censusOrigin, votes: choices, censusProof, processId, walletOrSigner: wallet, processKeys }) :
-      Voting.packageSignedEnvelope({ censusOrigin: processParams.censusOrigin, votes: choices, censusProof, processId, walletOrSigner: wallet })
+    // Prepare ZK Proof
+    const nullifier = Voting.getAnonymousVoteNullifier(account.secretKey, processId)
+    const { votePackage, keyIndexes } = Voting.packageVoteContent(choices, processKeys)
 
-    await VotingApi.submitEnvelope(envelope, wallet, gwPool)
+    const inputs: ZkInputs = {
+      censusRoot: processParams.rollingCensusRoot,
+      censusSiblings: censusProof.siblings,
+      maxSize: circuitInfo.maxSize,
+      keyIndex: censusProof.index,
+      nullifier,
+      secretKey: account.secretKey,
+      processId: Voting.getSnarkProcessId(processId),
+      votePackage
+    }
 
-    // Wait a bit
+    const zkProof = await ZkSnarks.computeProof(inputs, witnessGeneratorWasm, zKey)
+
+    const verifyProof = await ZkSnarks.verifyProof(JSON.parse(Buffer.from(vKey).toString()), zkProof.publicSignals as any, zkProof.proof as any)
+    if (config.stopOnError) assert(verifyProof)
+
+    // the next line is to place the 'nullifier' into the position 0 of
+    // publicSignals array. This is because in the Vochain, the rest of
+    // publicSignals from the array are not going to be used, as those inputs
+    // are going to be computed from known info of the process by the Vochain
+    // itself. And the Vochain only needs the 'nullifier' being in the first
+    // position of the array. In fact, the array could contain only 1 element
+    // (the 'nullifier').
+    zkProof.publicSignals = [nullifier.toString()]
+    // alternatively:
+    // zkProof.publicSignals[0] = zkProof.publicSignals[3];
+
+    const envelopeParams: AnonymousEnvelopeParams = {
+      votePackage,
+      processId,
+      zkProof,
+      nullifier,
+      circuitIndex: circuitInfo.index,
+      encryptionKeyIndexes: keyIndexes
+    }
+
+    // Package and submit
+    const envelope = Voting.packageAnonymousEnvelope(envelopeParams)
+
+    await VotingApi.submitEnvelope(envelope, null, gwPool)
+
     await new Promise(resolve => setTimeout(resolve, 11000))
 
-    // Check
-    const nullifier = Voting.getSignedVoteNullifier(wallet.address, processId)
+    // const nullifier = Voting.getAnonymousVoteNullifier(account.secretKey, processId)
     const { registered, date, block } = await VotingApi.getEnvelopeStatus(processId, nullifier, gwPool)
 
     if (config.stopOnError) assert(registered)
@@ -149,37 +218,37 @@ export async function checkVoteResults(processId: string, processParams: Process
   switch (config.votesPattern) {
     case "all-0":
       assert(rawResults.results[0].length >= 2)
-      assert.strictEqual(rawResults.results[0][0], config.numAccounts + "000000000000000000")
+      assert.strictEqual(rawResults.results[0][0], config.numAccounts.toString())
       assert.strictEqual(rawResults.results[0][1], "0")
       break
     case "all-1":
       assert(rawResults.results[0].length >= 2)
       assert.strictEqual(rawResults.results[0][0], "0")
-      assert.strictEqual(rawResults.results[0][1], config.numAccounts + "000000000000000000")
+      assert.strictEqual(rawResults.results[0][1], config.numAccounts.toString())
       break
     case "all-2":
       assert(rawResults.results[0].length >= 3)
       assert.strictEqual(rawResults.results[0][0], "0")
       assert.strictEqual(rawResults.results[0][1], "0")
-      assert.strictEqual(rawResults.results[0][2], config.numAccounts + "000000000000000000")
+      assert.strictEqual(rawResults.results[0][2], config.numAccounts.toString())
       break
     case "all-even":
       assert(rawResults.results[0].length >= 2)
       if (config.numAccounts % 2 == 0) {
-        assert.strictEqual(rawResults.results[0][0], (config.numAccounts / 2) + "000000000000000000")
-        assert.strictEqual(rawResults.results[0][1], (config.numAccounts / 2) + "000000000000000000")
+        assert.strictEqual(rawResults.results[0][0], (config.numAccounts / 2).toString())
+        assert.strictEqual(rawResults.results[0][1], (config.numAccounts / 2).toString())
       }
       else {
-        assert.strictEqual(rawResults.results[0][0], Math.ceil((config.numAccounts / 2)) + "000000000000000000")
-        assert.strictEqual(rawResults.results[0][1], Math.floor((config.numAccounts / 2)) + "000000000000000000")
+        assert.strictEqual(rawResults.results[0][0], (Math.ceil(config.numAccounts / 2)).toString())
+        assert.strictEqual(rawResults.results[0][1], (Math.floor(config.numAccounts / 2)).toString())
       }
       break
     case "incremental":
       assert.strictEqual(rawResults.results[0].length, 2)
       rawResults.results.forEach((question, i) => {
         for (let j = 0; j < question.length; j++) {
-          if (i == j) assert.strictEqual(question[j], config.numAccounts)
-          else assert.strictEqual(question[j], 0)
+          if (i == j) assert.strictEqual(question[j], config.numAccounts.toString())
+          else assert.strictEqual(question[j], "0")
         }
       })
       break