+++ packages/data-models/src/index.ts
@@ -1,5 +1,6 @@
 export * from './backup'
 export * from './entity'
 export * from './json-feed'
-export * from './process'
+export * from './voting-meta'
+export * from './voting-params'
 export * from './protobuf'