+++ packages/hashing/test/unit/poseidon.ts
@@ -7,6 +7,23 @@ import { Poseidon } from "../../src"
 addCompletionHooks()
 
 describe("Poseidon hashing", () => {
+    it("Should hash big int's", () => {
+        const items = [
+            {
+                input: BigInt("1"),
+                output: BigInt('18586133768512220936620570745912940619677854269274689475585506675881198879027')
+            },
+            {
+                input: BigInt("18586133768512220936620570745912940619677854269274689475585506675881198879027"),
+                output: BigInt('17744324452969507964952966931655538206777558023197549666337974697819074895989')
+            },
+        ]
+
+        items.forEach(item => {
+            expect(Poseidon.hash([item.input])).to.eq(item.output)
+        })
+    })
+
     it("Poseidon hashes should digest bigint arrays", () => {
         const BI_0 = BigInt("0")
         const BI_1 = BigInt("1")
@@ -36,5 +53,21 @@ describe("Poseidon hashing", () => {
 
         hash = Poseidon.hash([BI_1, BI_2, BI_3, BI_4, BI_5, BI_6])
         expect(hash.toString()).to.eq("20400040500897583745843009878988256314335038853985262692600694741116813247201")
+    })
+
+    it("Should fail when out of the field", () => {
+        const items = [
+            BigInt("-1"),
+            Poseidon.Q,
+            Poseidon.Q + BigInt("1"),
+        ]
+        items.forEach(item => {
+            try {
+                Poseidon.hash([item])
+            }
+            catch (err) {
+                expect(err.message).to.eq("One or more inputs are out of the Poseidon field")
+            }
+        })
     })
 })