+++ packages/voting/src/index.ts
@@ -1,2 +1,6 @@
 export * from "./voting"
-export * from "./entity"
+export * from "./voting-calls"
+export * from "./voting-oracle-calls"
+export * from "./entity-calls"
+
+export * from "./types"