+++ scripts/post-install.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+set -e
+
+for PKG in $(ls packages)
+do
+    echo "Running npm install on packages/$PKG"
+    cd packages/$PKG
+    npm install
+    cd -
+done